requirements.txt
flask==2.3.3
flask-limiter==3.5.0
flask-cors==4.0.0
flask-socketio==5.3.6
pyjwt==2.8.0
tronpy==0.4.2
websocket-client==1.6.4
requests==2.31.0
pycryptodome==3.20.0
eth-abi==4.0.0
schedule==1.2.2
ipfshttpclient==0.7.0
redis==6.2.0
gunicorn==21.2.0
web3==6.11.0







